Stack

+++ python/Q853_car_fleet.py
@@ -10,7 +10,7 @@ class Solution:
     def carFleet(self, target: int, position: List[int], speed: List[int]) -> int:
         count: int = 0
         previous_time: float = 0.0
-        required_time: list[int] = [
+        required_time: list[float] = [
             (target - p) / s for p, s in sorted(zip(position, speed), reverse=True)
         ]
